refactor(scaffolder): move getOctokitOptions to separate util file
