+++ plugins/scaffolder-backend-module-github/src/actions/githubActionsDispatch.ts
@@ -24,7 +24,7 @@ import {
   parseRepoUrl,
 } from '@backstage/plugin-scaffolder-node';
 import { Octokit } from 'octokit';
-import { getOctokitOptions } from './helpers';
+import { getOctokitOptions } from '../util';
 import { examples } from './githubActionsDispatch.examples';
 
 /**
