+++ plugins/scaffolder-backend-module-github/src/actions/githubIssuesLabel.examples.test.ts
@@ -25,7 +25,7 @@ import {
 import { createGithubIssuesLabelAction } from './githubIssuesLabel';
 import yaml from 'yaml';
 import { examples } from './githubIssuesLabel.examples';
-import { getOctokitOptions } from './helpers';
+import { getOctokitOptions } from '../util';
 
 jest.mock('./helpers', () => {
   return {
