+++ plugins/scaffolder-backend-module-github/src/actions/helpers.ts
@@ -15,13 +15,7 @@
  */
 
 import { Config } from '@backstage/config';
-import { assertError, InputError, NotFoundError } from '@backstage/errors';
-import {
-  DefaultGithubCredentialsProvider,
-  GithubCredentialsProvider,
-  ScmIntegrationRegistry,
-} from '@backstage/integration';
-import { OctokitOptions } from '@octokit/core/dist-types/types';
+import { assertError, NotFoundError } from '@backstage/errors';
 import { Octokit } from 'octokit';
 
 import {
@@ -35,75 +29,6 @@ import {
   entityRefToName,
 } from './gitHelpers';
 import { LoggerService } from '@backstage/backend-plugin-api';
-
-const DEFAULT_TIMEOUT_MS = 60_000;
-
-/**
- * Helper for generating octokit configuration options for given repoUrl.
- * If no token is provided, it will attempt to get a token from the credentials provider.
- * @public
- */
-export async function getOctokitOptions(options: {
-  integrations: ScmIntegrationRegistry;
-  credentialsProvider?: GithubCredentialsProvider;
-  host: string;
-  owner?: string;
-  repo?: string;
-  token?: string;
-}): Promise<OctokitOptions> {
-  const { integrations, credentialsProvider, host, owner, repo, token } =
-    options;
-
-  const requestOptions = {
-    // set timeout to 60 seconds
-    timeout: DEFAULT_TIMEOUT_MS,
-  };
-
-  const integrationConfig = integrations.github.byHost(host)?.config;
-
-  if (!integrationConfig) {
-    throw new InputError(`No integration for host ${host}`);
-  }
-
-  // short circuit the `githubCredentialsProvider` if there is a token provided by the caller already
-  if (token) {
-    return {
-      auth: token,
-      baseUrl: integrationConfig.apiBaseUrl,
-      previews: ['nebula-preview'],
-      request: requestOptions,
-    };
-  }
-
-  if (!owner || !repo) {
-    throw new InputError(
-      `No owner and/or owner provided, which is required if a token is not provided`,
-    );
-  }
-
-  const githubCredentialsProvider =
-    credentialsProvider ??
-    DefaultGithubCredentialsProvider.fromIntegrations(integrations);
-
-  const { token: credentialProviderToken } =
-    await githubCredentialsProvider.getCredentials({
-      url: `https://${host}/${encodeURIComponent(owner)}/${encodeURIComponent(
-        repo,
-      )}`,
-    });
-
-  if (!credentialProviderToken) {
-    throw new InputError(
-      `No token available for host: ${host}, with owner ${owner}, and repo ${repo}. Make sure GitHub auth is configured correctly. See https://backstage.io/docs/auth/github/provider for more details.`,
-    );
-  }
-
-  return {
-    auth: credentialProviderToken,
-    baseUrl: integrationConfig.apiBaseUrl,
-    previews: ['nebula-preview'],
-  };
-}
 
 export async function createGithubRepoWithCollaboratorsAndTopics(
   client: Octokit,
