+++ plugins/scaffolder-backend-module-github/src/actions/index.ts
@@ -29,5 +29,3 @@ export { createPublishGithubAction } from './github';
 export { createGithubAutolinksAction } from './githubAutolinks';
 export { createGithubPagesEnableAction } from './githubPagesEnable';
 export { createGithubBranchProtectionAction } from './githubBranchProtection';
-
-export { getOctokitOptions } from './helpers';
